Simplify AsyncFlowDefRunner
